feat!: implemented norm_int functions and CohSum
This feature is designed to allow for the correct implementation of
normalization integrals. Thanks to Lawrence Ng for pointing this out to
me.

BREAKING: This change requires AmpOps to be converted into CohSums, and
it changes the implementation of Model to load a Vec<CohSum> rather than
a single AmpOp.