feat!: Restructures AmpOp into concrete types
This commit removes AmpOp and replaces it with a trait, AmpLike, as well
as several concrete types that implement AmpLike: Amplitude, Real, Imag,
and Product. The amplitude! macro is removed, since it no longer has any
functionality beyond creating new Amplitudes, and that process is also
streamlined since we don't need to wrap everything in an enum.

This is a major performance increase. Benchmarks indicate upwards of a
>2x speed boost. The reason for this is because previously the compute
functions for AmpOps had a match statement, which doesn't play nice
inside the loop over all events. Refactoring this makes that match go
away, but at the cost of having to implement a bunch of traits for
operator overloading. This is now accomplished with a set of macros.

Future plans for more performance include getting rid of the initial
allocation of the cache vector in compute processes, but this might
change the signature to a &mut self, which I'd rather try to avoid if I
can. Also, I need to fix print_tree, it kind of works but it's a bit
messed up.

BREAKING: This commit removes AmpOp and replaces it with the AmpLike
trait. This refactor means developers should provide methods which give
their Nodes as Amplitudes rather than AmpOps, and the python interface
changes slightly (but you probably won't notice it). It's possible I've
missed some operator overloads in the process, so it might turn out that
some weird combination of them will not compile, but this can be fixed
quickly by updating the macros.