refactor: remove print_tree and replace it with Debug
When running in a Jupyter notebook, print_tree commands would print both in the notebook window and in the terminal. This is not exactly desired, and having printing commands lying around isn't that nice either in my opinion. I refactored print_tree-like statements into a trait AsTree which can be implemented on any object, not just AmpLikes, so CohSum and Model are included. All Displays will print as trees, while Debugs will print as one-liners with all info about the sub-objects as well (this might change in the future). In Python, __str__ will Display (print the tree) while __repr__ will Debug (print a one-liner).