feat: Add the ability to create new Nodes in Python
This commit creates a set of Python objects that can be used to create new Nodes on the Rust side, with the downside of not being able to use Rust's parallelism due to the GIL. To create a new Node in Python, one must first make a class which inherits from PyNode. This is an abstract base class with the requisite abstract methods to ensure the Node is valid. This class should then be wrapped in a Node class which can be input into the new Amplitude constructor (it also has an into_amplitude(name) method).