fix: update ELL evaluate methods to use the sum of weights rather than number of events and add normalization to intensity functions
I believe this is the correct way to do it based on some tests, although swapping acc for gen MC might screw something up still, I don't know. The proportionality constant in evaluate should be sum(data_weights)/sum(mc_weights), and when we want to evaluate a result over a new dataset, we must remember to scale by the correct weighted sizes of the datasets in mind as well as any event weights the input dataset might have associated with it. In the case where we project onto generated MC, these will just be 1, but it is possible to have non-unity weights in accepted MC.