refactor: Major update which implements generics
Replaces crate::Field with Field trait, a trait which implements
everything required by a Real numeric field in rustitude.

Furthermore replaces every struct with a generic struct over Field. This
is a major refactor since it will require all downstream amplitudes to
be updated entirely. The benefit is that f64 and f32 versions of the
crate are simultaneously supported without a feature gate. This allows
downstream crates like the py-rustitude Python bindings to use both
precisions right from the package without having to recompile from a git
repo.