feat: adds f64 and f32 variants of each Python binding
This commit takes advantage of the generics in rustitude-core to create
optional f64 and f32 versions of each PyO3-class struct. These are
implemented with aliases to the old class names for the f64 versions, so
old code should not break under these changes unless rustitude was built
with the `float` feature. Also adds a `f32` boolean flag to the `open`
command to allow it to automatically create `DatasetF32`s.