feat!: removed CohSum and replaced with Sum + NormSqr
BREAKING CHANGE: This commit will break basically everything, but again, it's probably for the best. This allows for much cleaner and more concise model construction. For instance, we can now take the real part of a sum of terms. This also adds some missing impl Ops methods, so we should be able to arbitrarily add and multiply sums, products, and other terms. Multiplying sums with sums is still unimplemented, but this commit allows for it without a breaking change if it is desired in the future. This also makes it much clearer to end-users what the `+` operator does to amplitudes, and separates the coherent sum functionality to be an implicit part of a model rather than some explicit part of amplitude construction. This also removes the need for conversions to CohSum before model construction, so `model!(a_single_term, a_sum_of_terms)` works now! The downside is that inputs to `Model::new` have to be boxed, but the `model!` macro makes the syntax convenient.