fix: place `Model` amplitudes under a `Arc<RwLock<Amplitude<F>>>` to avoid duplication
In the past, I moved away from this because locking was a hit to performance. However, it turns out that with minimal modification, I can lock the amplitudes ahead of the loop over events, so they only get locked once rather than per event. This makes the performance fairly comparable, and makes `Model`s clone-able without duplicating all of the precomputed data. This saves a huge amount of memory when an optimizer clones an ExtendedLogLikelihood, since it's basically a stack copy now instead of a heap clone.