feat: add `Decay` enum to basically every amplitude to avoid hardcoding particle indices
This allows for expansion in the future if we want to implement ways to combine groups of particles into a decaying particle, but for now just covers the basics of two- and three-body decays involving particles already in the list of daughter_p4s. The Python API has been updated in a breaking way, but default kwargs remain the same.

BREAKING CHANGE: This fundamentally changes the way GlueX amplitudes are formed in both rust and python. While most defaults remain unchanged, using non-default kwargs will require code to change!