fix: resolves a huge bug I introduced when I put the `Model` `amplitudes` field under an `Arc`
Previously, creating two `Manager`s with the same model wouldn't actually duplicate the model properly, the underlying amplitudes would still be the same references, so if you precalculate things with one dataset followed by another, the second will overwrite the first.

This also adds a test to make sure this doesn't happen again, as well as some convenient testing functions and a way to reset the indices of a `Dataset`.