feat!: refine the `Field` trait by removing the `RealField` subtrait
Additionally, I have removed the f-prefixed mathematical functions, since the only conflicts were with the `nalgebra` implementations. From now on, we will prefer using `num` traits and building in additional functionality. This also adds several convenience methods for getting the magnitude of EPS and normalized `Vector3` fields using the new `unit` method to avoid requiring `RealField` for this very simple operation. Additionally, several macros are added to make conversion from raw values easier. In the future, trait implementations might be added as well to provide shortcuts to these methods, but that is a WIP.